+++ src/app/calendar/page.tsx
@@ -6,52 +6,21 @@ import FilterBtn from '@/components/calender/FilterBtn'
 import AuthModal from '@/components/common/AuthModal'
 import BookmarkModal from '@/components/common/BookmarkModal'
 import { CATEGORIES } from '@/constants/category'
-import { useEffect, useState } from 'react'
+import { useCalendar } from '@/hooks/calendar/useCalendar'
 
 export default function Page() {
-  const [showModal, setShowModal] = useState<boolean>(false)
-  const [selectedCategories, setSelectedCategories] = useState<string[]>(
-    CATEGORIES.map((category) => category.value),
-  )
-  const handleOpenModal = () => {
-    fetchUserProfile()
-  }
+  const {
+    showModal,
+    authModalOpen,
+    modalOpen,
+    selectedCategories,
+    handleOpenModal,
+    handleCloseModal,
+    handleFilter,
+    setAuthModalOpen,
+    setModalOpen,
+  } = useCalendar()
 
-  const handleCloseModal = () => setShowModal(false)
-  const [authModalOpen, setAuthModalOpen] = useState(false)
-  const handleFilter = (category: string) => {
-    setSelectedCategories((prev) => {
-      if (prev.includes(category)) {
-        const newCategories = prev.filter((cat) => cat !== category)
-        return newCategories.length === 0
-          ? CATEGORIES.map((cat) => cat.value)
-          : newCategories
-      } else {
-        return [...prev, category]
-      }
-    })
-  }
-  const fetchUserProfile = async () => {
-    try {
-      const response = await fetch('/api/v1/users', {
-        method: 'GET',
-        credentials: 'include',
-      })
-      if (response.status === 401) {
-        // 401이면 로그인 모달 오픈
-        setAuthModalOpen(true)
-        return
-      }
-      setShowModal(true)
-      if (!response.ok) {
-        throw new Error('유저조회 실패')
-      }
-    } catch (err: any) {
-      setAuthModalOpen(true)
-      throw new Error(err)
-    }
-  }
-  const [modalOpen, setModalOpen] = useState(false)
   return (
     <div className="flex flex-col items-center">
       <AuthModal
